Try hand implementing the lexer
String literals were a pain to work with. I ended up using `{` as the
interpolation delimiter instead of `#{` because the regex got too
unwieldy with needing to exclude a two character symbol.

The string stuff is definitely easier but overall it's a lot more code
and less clear what's going on. Perhaps there's a way to primarily use
leex, but escape into elixir when needed...